refactor: refactor code
